Synth xbox (#34)
* filling out some more stuff mainly in meta

focused on meta, filled out a bunch of functions. also re-worked ones I basically copied and pasted from rb3 to the proper adresses for dc3!

* Update SongMgr.cpp

* forgot some functions in SongMgr

that really dumb last commit made me realize not every function in SongMgr was in the header/cpp file so i took care of that

* formatting

formatted all the files ive worked on

* Header Virtual Functions

---------

Co-authored-by: Iman Din <38789931+imandin408@users.noreply.github.com>